python-ml/requirements.txt
flask==3.0.0
flask-cors==4.0.0
opencv-python==4.8.1.78
scikit-learn==1.3.2
scikit-image==0.22.0
joblib==1.3.2
pytesseract==0.3.10
pillow==10.1.0
numpy==1.26.2
